% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{pandoc_latex_template}
\alias{pandoc_latex_template}
\title{R Markdown output format for pandoc-latex-template report}
\usage{
pandoc_latex_template(dev = "pdf",
  template = template_resources("pandoc-latex-template", "eisvogel.tex"),
  ...)
}
\arguments{
\item{..., keep_tex, latex_engine, citation_package, highlight, fig_caption, md_extensions}{Arguments passed to \code{rmarkdown::\link{pdf_document}()}.}
}
\value{
An R Markdown output format.
}
\description{
Formats are based on \code{rmarkdown::pdf_document()}, with a
custom Pandoc LaTeX template and different default values for other arguments
(e.g., \code{keep_tex = TRUE}).
}
\section{Details}{
 You can find more details about each output format below.
}

\section{\code{pandoc_latex_template}}{
 Format for creating an document adapted from
  \url{https://github.com/Wandmalfarbe/pandoc-latex-template}.
}

\examples{
\dontrun{
rmarkdown::draft("MyArticle.Rmd", template = "pandoc_latex_template", package = "pandoclatextemplate")
}
}
